+++ signal_analog/util.py
@@ -90,3 +90,8 @@ def pp_json(dictionary):
 def empty_body():
     """Returns  an empty body when making requests to SignalFx."""
     return lambda x: None
+
+def snake_to_camel(s):
+    words = s.split('_')
+    head, rest = words[0], words[1:]
+    return head.lower() + ''.join(map(str.title, rest))